Modul_1(python_core)/self_study/companies_general_data_collection_full/requirements.txt
aiohttp==3.10.11
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
validators==0.34.0
email-validator==2.2.0
dnspython==2.6.1
tldextract==5.1.2
